+++ src/main/java/com/matyrobbrt/keybindbundles/render/KeybindSelectionScreen.java
@@ -75,9 +75,7 @@ public class KeybindSelectionScreen extends Screen {
         var key = KeyMappingUtil.getByName(mapping.bind.getEntries().get(selectionIndex).key());
         Minecraft.getInstance().setScreen(null);
         if (key != null) {
-            KeyMappingUtil.press(key);
-            KeyMappingUtil.click(key);
-            Minecraft.getInstance().execute(() -> KeyMappingUtil.release(key));
+            KeyMappingUtil.tap(key);
         }
     }
 
@@ -95,7 +93,7 @@ public class KeybindSelectionScreen extends Screen {
         GLFW.glfwGetCursorPos(mainWindow.getWindow(), xPos, yPos);
 
         double scaledX = xPos[0] - (windowWidth / 2.0f);
-        double scaledY = yPos[1] - (windowHeight / 2.0f);
+        double scaledY = yPos[0] - (windowHeight / 2.0f);
 
         double distance = Math.sqrt(scaledX * scaledX + scaledY * scaledY);
         double radius = RadialMenuRenderer.OUTER * ((double) windowWidth / mainWindow.getGuiScaledWidth()) * 1.1;